Trait pliantdb::connection::Connection

pub trait Connection<'a>: Send + Sync {
    fn collection<C: Collection + 'static>(
        &'a self
    ) -> Result<Collection<'a, Self, C>, Error>
    where
        Self: Sized;

#[must_use]    fn insert<'life0, 'async_trait, C: Collection>(
        &'life0 self, 
        contents: Vec<u8>
    ) -> Pin<Box<dyn Future<Output = Result<Header, Error>> + Send + 'async_trait>>
    where
        C: 'async_trait,
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn update<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        doc: &'life1 mut Document<'life2>
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn get<'life0, 'async_trait, C: Collection>(
        &'life0 self, 
        id: Uuid
    ) -> Pin<Box<dyn Future<Output = Result<Option<Document<'static>>, Error>> + Send + 'async_trait>>
    where
        C: 'async_trait,
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn apply_transaction<'life0, 'async_trait>(
        &'life0 self, 
        transaction: Transaction<'static>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<OperationResult>, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn list_executed_transactions<'life0, 'async_trait>(
        &'life0 self, 
        starting_id: Option<u64>, 
        result_limit: Option<u64>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Executed<'static>>, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

a trait that defines all interactions with a Database, regardless of whether it is local or remote

Required methods

fn collection<C: Collection + 'static>(
&'a self
) -> Result<Collection<'a, Self, C>, Error> where
Self: Sized,

access a collection for the connected Database

#[must_use]fn insert<'life0, 'async_trait, C: Collection>(
&'life0 self,
contents: Vec<u8>
) -> Pin<Box<dyn Future<Output = Result<Header, Error>> + Send + 'async_trait>> where
C: 'async_trait,
'life0: 'async_trait,
Self: 'async_trait,

insert a newly created document into the connected Database for the collection C

#[must_use]fn update<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
doc: &'life1 mut Document<'life2>
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

update an existing document in the connected Database for the collection C. Upon success, doc.revision will be updated with the new revision.

#[must_use]fn get<'life0, 'async_trait, C: Collection>(
&'life0 self,
id: Uuid
) -> Pin<Box<dyn Future<Output = Result<Option<Document<'static>>, Error>> + Send + 'async_trait>> where
C: 'async_trait,
'life0: 'async_trait,
Self: 'async_trait,

retrieve a stored document from collection C identified by id

#[must_use]fn apply_transaction<'life0, 'async_trait>(
&'life0 self,
transaction: Transaction<'static>
) -> Pin<Box<dyn Future<Output = Result<Vec<OperationResult>, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

apply a transaction to the database. If any operation in the transaction fails, none of the operations will be applied to the database.

#[must_use]fn list_executed_transactions<'life0, 'async_trait>(
&'life0 self,
starting_id: Option<u64>,
result_limit: Option<u64>
) -> Pin<Box<dyn Future<Output = Result<Vec<Executed<'static>>, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

list executed transactions from this database. By default, a maximum of 1000 entries will be returned, but that limit can be overridden by setting result_limit. A hard limit of 100,000 results will be returned. To begin listing after another known transaction_id, pass transaction_id + 1 into starting_id.

Loading content...