Trait pliantdb::core::backend::CustomApi

pub trait CustomApi: Debug + Send + Sync {
    type Request: Serialize + for<'de> Deserialize<'de> + Send + Sync + Debug;
    type Response: Clone + Serialize + for<'de> Deserialize<'de> + Send + Sync + Debug;
    type Dispatcher: Dispatcher<Self::Request> + Debug;
}

A definition of a custom API. The Request associated type is what a connected client can send. The Response associated type is what your API is expecetd to send from the server to the client. The Dispatcher is the type that will handle the custom API requests.

This feature is powered by [actionable] through these derives:

Action (trait and derive macro)

Dispatcher (trait and derive macro)

Actionable (derive macro)

Associated Types

type Request: Serialize + for<'de> Deserialize<'de> + Send + Sync + Debug

The type that represents an API request. This type is what clients will send to the server.

type Response: Clone + Serialize + for<'de> Deserialize<'de> + Send + Sync + Debug

The type that represents an API response. This type will be sent to clients from the server.

type Dispatcher: Dispatcher<Self::Request> + Debug

The type that implements the Dispatcher trait.

Loading content...