Trait pliantdb::local::core::connection::Connection

pub trait Connection<'a>: Send + Sync {
    pub fn collection<C>(&'a self) -> Result<Collection<'a, Self, C>, Error>
    where
        C: 'static + Collection;

#[must_use]    pub fn get<'life0, 'async_trait, C>(
        &'life0 self, 
        id: u64
    ) -> Pin<Box<dyn Future<Output = Result<Option<Document<'static>>, Error>> + 'async_trait + Send, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait,
        C: Collection + 'async_trait;

#[must_use]
#[must_use]    pub fn query<'k, 'life0, 'async_trait, V>(
        &'life0 self, 
        query: View<'a, Self, V>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Serialized, Global>, Error>> + 'async_trait + Send, Global>>
    where
        'k: 'async_trait,
        'a: 'async_trait,
        'life0: 'async_trait,
        Self: 'async_trait,
        V: View + 'async_trait;

#[must_use]
#[must_use]    pub fn reduce<'k, 'life0, 'async_trait, V>(
        &'life0 self, 
        query: View<'a, Self, V>
    ) -> Pin<Box<dyn Future<Output = Result<<V as View>::Value, Error>> + 'async_trait + Send, Global>>
    where
        'k: 'async_trait,
        'a: 'async_trait,
        'life0: 'async_trait,
        Self: 'async_trait,
        V: View + 'async_trait;

#[must_use]    pub fn apply_transaction<'life0, 'async_trait>(
        &'life0 self, 
        transaction: Transaction<'static>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<OperationResult, Global>, Error>> + 'async_trait + Send, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    pub fn list_executed_transactions<'life0, 'async_trait>(
        &'life0 self, 
        starting_id: Option<u64>, 
        result_limit: Option<usize>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Executed<'static>, Global>, Error>> + 'async_trait + Send, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    pub fn insert<'life0, 'async_trait, C>(
        &'life0 self, 
        contents: Vec<u8, Global>
    ) -> Pin<Box<dyn Future<Output = Result<Header, Error>> + 'async_trait + Send, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait,
        C: Collection + 'async_trait,
    { ... }

#[must_use]    pub fn update<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        doc: &'life1 mut Document<'life2>
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 'async_trait + Send, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait,
    { ... }

#[must_use]    pub fn delete<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        doc: &'life1 Document<'life2>
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 'async_trait + Send, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait,
    { ... }

#[must_use]    pub fn view<V>(&'a self) -> View<'a, Self, V>
    where
        V: View,
    { ... }
}

Defines all interactions with a schema::Database, regardless of whether it is local or remote.

Required methods

pub fn collection<C>(&'a self) -> Result<Collection<'a, Self, C>, Error> where
C: 'static + Collection,

Accesses a collection for the connected schema::Database.

#[must_use]pub fn get<'life0, 'async_trait, C>(
&'life0 self,
id: u64
) -> Pin<Box<dyn Future<Output = Result<Option<Document<'static>>, Error>> + 'async_trait + Send, Global>> where
'life0: 'async_trait,
Self: 'async_trait,
C: Collection + 'async_trait,

Retrieves a stored document from Collection C identified by id.

#[must_use] #[must_use]pub fn query<'k, 'life0, 'async_trait, V>(
&'life0 self,
query: View<'a, Self, V>
) -> Pin<Box<dyn Future<Output = Result<Vec<Serialized, Global>, Error>> + 'async_trait + Send, Global>> where
'k: 'async_trait,
'a: 'async_trait,
'life0: 'async_trait,
Self: 'async_trait,
V: View + 'async_trait,

Queries for view entries matching [ViewQuery].

#[must_use] #[must_use]pub fn reduce<'k, 'life0, 'async_trait, V>(
&'life0 self,
query: View<'a, Self, V>
) -> Pin<Box<dyn Future<Output = Result<<V as View>::Value, Error>> + 'async_trait + Send, Global>> where
'k: 'async_trait,
'a: 'async_trait,
'life0: 'async_trait,
Self: 'async_trait,
V: View + 'async_trait,

Reduces the view entries matching [ViewQuery].

#[must_use]pub fn apply_transaction<'life0, 'async_trait>(
&'life0 self,
transaction: Transaction<'static>
) -> Pin<Box<dyn Future<Output = Result<Vec<OperationResult, Global>, Error>> + 'async_trait + Send, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Applies a Transaction to the schema::Database. If any operation in the Transaction fails, none of the operations will be applied to the schema::Database.

#[must_use]pub fn list_executed_transactions<'life0, 'async_trait>(
&'life0 self,
starting_id: Option<u64>,
result_limit: Option<usize>
) -> Pin<Box<dyn Future<Output = Result<Vec<Executed<'static>, Global>, Error>> + 'async_trait + Send, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Lists executed Transactions from this schema::Database. By default, a maximum of 1000 entries will be returned, but that limit can be overridden by setting result_limit. A hard limit of 100,000 results will be returned. To begin listing after another known transaction_id, pass transaction_id + 1 into starting_id.

Loading content...
Provided methods

#[must_use]pub fn insert<'life0, 'async_trait, C>(
&'life0 self,
contents: Vec<u8, Global>
) -> Pin<Box<dyn Future<Output = Result<Header, Error>> + 'async_trait + Send, Global>> where
'life0: 'async_trait,
Self: 'async_trait,
C: Collection + 'async_trait,

Inserts a newly created document into the connected schema::Database for the Collection C.

#[must_use]pub fn update<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
doc: &'life1 mut Document<'life2>
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 'async_trait + Send, Global>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Updates an existing document in the connected schema::Database for the Collection C. Upon success, doc.revision will be updated with the new revision.

#[must_use]pub fn delete<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
doc: &'life1 Document<'life2>
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + 'async_trait + Send, Global>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Removes a Document from the database.

#[must_use]pub fn view<V>(&'a self) -> View<'a, Self, V> where
V: View,

Initializes [ViewQuery] for schema::View V.

Loading content...