Trait pliantdb::local::core::schema::View

pub trait View: 'static + Send + Sync + Debug {
    type Collection: Collection;
    type MapKey: 'static + Key;
    type MapValue: for<'de> Deserialize<'de> + Serialize;
    type Reduce: for<'de> Deserialize<'de> + Serialize;
    pub fn version(&self) -> u64;

    pub fn name(&self) -> Cow<'static, str>;

    pub fn map(
        &self, 
        document: &Document<'_>
    ) -> Result<Option<Map<Self::MapKey, Self::MapValue>>, Error>;

    pub fn reduce(
        &self, 
        mappings: &[Map<Self::MapKey, Self::MapValue>], 
        rereduce: bool
    ) -> Result<Self::Reduce, Error> { ... }
}

A map/reduce powered indexing and aggregation schema.

Inspired by CouchDB's view system

This implementation is under active development, our own docs explaining our implementation will be written as things are solidified.

Associated Types

type Collection: Collection

The collection this view belongs to

type MapKey: 'static + Key

The key for this view.

type MapValue: for<'de> Deserialize<'de> + Serialize

An associated type that can be stored with each entry in the view.

type Reduce: for<'de> Deserialize<'de> + Serialize

When implementing reduce(), this is the returned type.

Loading content...
Required methods

pub fn version(&self) -> u64

The version of the view. Changing this value will cause indexes to be rebuilt.

pub fn name(&self) -> Cow<'static, str>

The name of the view. Must be unique per collection.

pub fn map(
&self,
document: &Document<'_>
) -> Result<Option<Map<Self::MapKey, Self::MapValue>>, Error>

The map function for this view. This function is responsible for emitting entries for any documents that should be contained in this View. If None is returned, the View will not include the document.

Loading content...
Provided methods

pub fn reduce(
&self,
mappings: &[Map<Self::MapKey, Self::MapValue>],
rereduce: bool
) -> Result<Self::Reduce, Error>

The reduce function for this view. If Err(Error::ReduceUnimplemented) is returned, queries that ask for a reduce operation will return an error. See CouchDB's Reduce/Rereduce documentation for the design this implementation will be inspired by

Loading content...