Module pliantdb::server::local::core

Core functionality and types for PliantDB.

Modules

circulate
Lightweight async PubSub framework.

connection
Types for interacting with PliantDB.

document
Types for interacting with Documents.

kv
Types for utilizing a lightweight atomic Key-Value store.

limits
Limits used within PliantDB.

networking
Types for implementing the PliantDB network protocol.

num_traits
Numeric traits for generic mathematics

pubsub
Types for Publish/Subscribe (PubSub) messaging.

schema
Types for defining database schema.

test_util
Shared schemas and utilities used for unit testing.

transaction
Types for executing transactions.

Macros

define_connection_test_suite
Creates a test suite that tests methods available on Connection

define_kv_test_suite
Defines the Kv test suite

define_pubsub_test_suite
Expands into a suite of pubsub unit tests using the passed type as the test harness.

Enums

Error
an enumeration of errors that this crate can produce