Trait pliantdb::server::local::core::networking::fabruic::dangerous::KeyPair

pub trait KeyPair {
    pub fn serialize<S>(
        key_pair: &Self, 
        serializer: S
    ) -> Result<<S as Serializer>::Ok, <S as Serializer>::Error>
    where
        S: Serializer;
}

Gives serialization access to KeyPair.

Security

This is only dangerous in the sense that you aren’t supposed to leak the PrivateKey. Make sure to use this carefully!

Required methods

pub fn serialize<S>(
key_pair: &Self,
serializer: S
) -> Result<<S as Serializer>::Ok, <S as Serializer>::Error> where
S: Serializer,

Serialize with [serde].

Security

This is only dangerous in the sense that you aren’t supposed to leak the PrivateKey. Make sure to use this carefully!

Errors

S::Error if serialization failed.

Loading content...