Module pliantdb::server::local::core::permissions

Types for creating and validating permissions.

Modules

pliant
Types used for granting permissions within PliantDb.

Structs

ActionName
A unique name of an action.

PermissionDenied
An action was denied.

Permissions
A collection of allowed permissions. This is constructed from a Vec<Statement>. By default, no actions are allowed on any resources.

ResourceName
A unique name/identifier of a resource.

Statement
A statement of permissions. A statement describes whether one or more actions should be allowed to be taken against resources.

Enums

ActionNameList
A list of ActionNames.

Identifier
A single element of a ResourceName

Traits

Action
An action that can be allowed or disallowed.

Dispatcher
Dispatches T to an appropriate handler. This trait is derivable.

Derive Macros

Action
Derives the actionable::Action trait.

Actionable
Derives a set of traits that can be used to implement a permissions-driven API. There are options that can be customized with the #[actionable] attribute at the enum level:

Dispatcher
Derives the Dispatcher trait.