Module pliantdb::server::local::core::permissions::pliant

Types used for granting permissions within PliantDb.

Enums

DatabaseAction
Actions that operate on a specific database.

DocumentAction
Actions that operate on a document.

EncryptionKeyAction
Actions that use encryption keys.

KvAction
Actions that operate on the key-value store.

PliantAction
Actions that can be permitted within PliantDb.

PubSubAction
Actions that operate on the PubSub system.

ServerAction
Actions that operate on a server.

TransactionAction
Actions that operate on transactions.

ViewAction
Actions that operate on a view.

Functions

collection_resource_name
Creates a resource name for a collection within a database.

database_resource_name
Creates a resource name with the database name.

document_resource_name
Creates a resource name for a document id within collection within database.

encryption_key_resource_name
Creates a resource name for encryption key key_id.

kv_key_resource_name
Creates a resource name for key within namespace within the key-value store of database.

pliantdb_resource_name
The base PliantDb resource namespace. All database objects have this as their first name segment.

pubsub_topic_resource_name
Creates a resource name for PubSub topic within database.

user_resource_name
Creates a resource name for the user with username.

view_resource_name
Creaets a resource name for a view within database.