Module pliantdb::server::local::core::schema::view::map

Types for defining a Map within a View.

Structs

Map
A document’s entry in a View’s mappings.

MappedDocument
A document’s entry in a View’s mappings.

MappedValue
A key value pair

Serialized
Represents a document’s entry in a View’s mappings, serialized and ready to store.

Traits

EnumKey
Adds Key support to an enum. Requires implementing ToPrimitive and FromPrimitive, or using a crate like num-derive to do it automatically. Take care when using enums as keys: if the order changes or if the meaning of existing numerical values changes, make sure to update any related views’ version number to ensure the values are re-evaluated.

Key
A trait that enables a type to convert itself to a big-endian/network byte order.